+++ src/templates/global_fns.rs
@@ -4,6 +4,7 @@ use std::path::{PathBuf};
 use tera::{GlobalFn, Value, from_value, to_value, Result};
 
 use content::Page;
+use site::resolve_internal_link;
 
 
 pub fn make_get_page(all_pages: &HashMap<PathBuf, Page>) -> GlobalFn {
@@ -24,6 +25,21 @@ pub fn make_get_page(all_pages: &HashMap<PathBuf, Page>) -> GlobalFn {
                 Err(_) => Err(format!("`get_page` received path={:?} but it requires a string", val).into()),
             },
             None => Err("`get_page` requires a `path` argument.".into()),
+        }
+    })
+}
+
+pub fn make_get_url(permalinks: HashMap<String, String>,) -> GlobalFn {
+    Box::new(move |args| -> Result<Value> {
+        match args.get("link") {
+            Some(val) => match from_value::<String>(val.clone()) {
+                Ok(v) => match resolve_internal_link(&v, &permalinks) {
+                    Ok(url) => Ok(to_value(url).unwrap()),
+                    Err(_) => Err(format!("Could not resolve URL for link `{}` not found.", v).into())
+                },
+                Err(_) => Err(format!("`get_url` received link={:?} but it requires a string", val).into()),
+            },
+            None => Err("`get_url` requires a `link` argument.".into()),
         }
     })
 }