+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix missing serialized data for sections
 - Change the single item template context for categories/tags
+- Add a `get_url` global Tera function
 
 ## 0.0.5 (2017-05-15)